+++ CHANGELOG.md
@@ -13,7 +13,7 @@ This release significantly revises the extraction mechanism, which means we can 
 - Podman support
 - System: uClibc
 - Solver collection: SAT heritage
-- Solver: IsaSAT
+- Solvers: IsaSAT, MergeSat
 - Resumable stages for extraction, transformation, and solving
 - Flexible solver queries for void, backbone, and partial configuration analysis